[SRDF] Support SRDF as XML string and split declaration and definition
